{"schema":"decodex/commit/1","summary":"Move overlay config defaults into session contracts","authority":"manual"}
